net/mrvl: fix port id type
port id should be defined as uint16_t.

Fixes: f8244c6399d9 ("ethdev: increase port id range")

+++ drivers/net/mrvl/mrvl_qos.c
@@ -486,7 +486,7 @@ setup_tc(struct pp2_ppio_tc_params *param, uint8_t inqs,
  * @returns 0 in case of success, negative value otherwise.
  */
 int
-mrvl_configure_rxqs(struct mrvl_priv *priv, uint8_t portid,
+mrvl_configure_rxqs(struct mrvl_priv *priv, uint16_t portid,
 	uint16_t max_queues)
 {
 	size_t i, tc;
